requirements.txt
streamlit==1.35.0
torch==2.3.0
torchvision==0.18.0
pillow==10.4.0